requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==0.28.1
pydantic==2.4.2
httpx==0.24.1
requests==2.31.0